Do a cast to make llvm build it: /usr/src/usr.sbin/ndmpd/../../contrib/ndmpd/src/ndmpd_func.c:1084:23: error: comparison of       integers of different signs: 'unsigned int' and 'int' [-Werror,-Wsign-compare]         if ((message & 0xff) >= ni->hd_cnt)              ~~~~~~~~~~~~~~  ^  ~~~~~~~~~~

+++ src/ndmpd_func.c
@@ -1080,8 +1080,7 @@ ndmp_get_interface(ndmp_message message)
 {
 	ndmp_handler_t *ni = &ndmp_msghdl_tab[(message >> 8) % INT_MAXCMD];
 
-
-	if ((message & 0xff) >= ni->hd_cnt)
+	if ((message & 0xff) >= (unsigned int)ni->hd_cnt)
 		return (NULL);
 
 	/* Sanity check */
